+++ Server/DealnetPortal.Api.Integration/Utility/XlsxExporter.cs
@@ -151,10 +151,9 @@ namespace DealnetPortal.Api.Integration.Utility
                                 LoanTerm = contract.Equipment.LoanTerm ?? 0,
                                 AmortizationTerm = contract.Equipment.AmortizationTerm ?? 0,
                                 PriceOfEquipment = (double)priceOfEquipment,
-                                //  EquipmentCashPrice = (double?)contract.Equipment?.NewEquipment.Sum(x => x.Cost) ?? 0,
-                                AdminFee = contract.Equipment.AdminFee ?? 0,
-                                DownPayment = contract.Equipment.DownPayment ?? 0,
-                                CustomerRate = contract.Equipment.CustomerRate ?? 0,
+                                AdminFee = (double)(contract.Equipment.AdminFee ?? 0),
+                                DownPayment = (double)(contract.Equipment.DownPayment ?? 0),
+                                CustomerRate = (double)(contract.Equipment.CustomerRate ?? 0),
                                 IsClarity = contract.Equipment.IsClarityProgram,
                                // IsOldClarityDeal = contract.Equipment?.IsClarityProgram == null && contract.Dealer?.Tier?.Name == _config.GetSetting(WebConfigKeys.CLARITY_TIER_NAME)
                             };
